+++ tests/testthat/test-proportional_hazards-survival.R
@@ -726,3 +726,39 @@ test_that("get_missings_coxph() can identify missings with two strata terms", {
   )
   expect_true(is.null(get_missings_coxph(f_fit$fit, na_0_data_x)))
 })
+
+
+# fit via matrix interface ------------------------------------------------
+
+test_that("`fix_xy()` works", {
+  lung_x <- as.matrix(lung[, c("age", "ph.ecog")])
+  lung_y <- Surv(lung$time, lung$status)
+  lung_pred <- lung[1:5, ]
+
+  spec <- proportional_hazards() %>% set_engine("survival")
+  f_fit <- fit(spec, Surv(time, status) ~ age + ph.ecog, data = lung)
+  xy_fit <- fit_xy(spec, x = lung_x, y = lung_y)
+
+  elements_to_ignore <- c("call", "formula", "terms", "model")
+  f_ignore <- which(names(f_fit$fit) %in% elements_to_ignore)
+  xy_ignore <- which(names(xy_fit$fit) %in% elements_to_ignore)
+  expect_equal(
+    f_fit$fit[-f_ignore],
+    xy_fit$fit[-xy_ignore],
+    ignore_formula_env = TRUE
+  )
+
+  f_pred_time <- predict(f_fit, new_data = lung_pred, type = "time")
+  xy_pred_time <- predict(xy_fit, new_data = lung_pred, type = "time")
+  expect_equal(f_pred_time, xy_pred_time)
+
+  f_pred_survival <- predict(f_fit, new_data = lung_pred,
+                             type = "survival", time = c(100, 200))
+  xy_pred_survival <- predict(xy_fit, new_data = lung_pred,
+                              type = "survival", time = c(100, 200))
+  expect_equal(f_pred_survival, xy_pred_survival)
+
+  f_pred_lp <- predict(f_fit, new_data = lung_pred, type = "linear_pred")
+  xy_pred_lp <- predict(xy_fit, new_data = lung_pred, type = "linear_pred")
+  expect_equal(f_pred_lp, xy_pred_lp)
+})